[#5] Restructure project layout
