+++ pkg/dunner/dunner.go
@@ -1,16 +1,16 @@
-package DunnerService
+package dunner
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/leopardslab/Dunner/src/services/DockerService"
-	"github.com/leopardslab/Dunner/src/services/ConfigService"
+	"github.com/leopardslab/Dunner/pkg/docker"
+	"github.com/leopardslab/Dunner/pkg/config"
 )
 
 func Do(cmd *cobra.Command, args []string) {
 
-	var configs = ConfigService.GetConfigs()
+	var configs = config.GetConfigs()
 	for _, stepDefinition := range configs.Tasks[args[0]] {
-		step := DockerService.Step {
+		step := docker.Step {
 			Task: args[0],
 			Name: stepDefinition.Name,
 			Image: stepDefinition.Image,
